simple-rag-chatbot-with-langgraph/requirements.txt
langchain
langchain-community
langchain_nvidia_ai_endpoints
langchain_openai
pypdf
langchain_experimental
langchain-qdrant
pytest
python-dotenv
gradio
